refactor(design): extracted design reducers & some of the export utils
